HF 30 - Private message groups - Members, Group cost chain properties, fixes

+++ plugins/private_message/private_message_evaluators.cpp
@@ -4,10 +4,17 @@
 #include <golos/plugins/private_message/private_message_evaluators.hpp>
 
 uint32_t MAX_GROUPS_LIMIT = 100; // per creator
-uint32_t MAX_ADMINS_LIMIT = 50; // per group
 uint32_t MAX_MODERS_LIMIT = 50; // per group
 
 namespace golos { namespace plugins { namespace private_message {
+
+void check_golos_power(const account_object& acc, const database& _db) {
+    auto min_gp = _db.get_min_gp_for_groups();
+    GOLOS_CHECK_LOGIC(acc.vesting_shares >= min_gp.first,
+        logic_errors::too_low_gp,
+        "Too low golos power: ${r} is required, ${p} is present.",
+        ("r", min_gp.second)("p", acc.vesting_shares));
+}
 
 struct private_message_extension_visitor {
     private_message_extension_visitor(const account_name_type& _requester,
@@ -33,15 +40,12 @@ struct private_message_extension_visitor {
         bool exists = pgm_itr != pgm_idx.end();
 
         if (exists) {
-            GOLOS_CHECK_LOGIC(pgm_itr->member_type != private_group_member_type::retired,
-                logic_errors::unauthorized, "You are retired.");
             GOLOS_CHECK_LOGIC(pgm_itr->member_type != private_group_member_type::banned,
                 logic_errors::unauthorized, "You are banned.");
         }
 
         bool is_owner = pgo.owner == requester;
-        bool is_admin = is_owner || (exists && pgm_itr->member_type == private_group_member_type::admin);
-        bool is_moder = is_admin || (exists && pgm_itr->member_type == private_group_member_type::moder); 
+        bool is_moder = is_owner || (exists && pgm_itr->member_type == private_group_member_type::moder); 
 
         if (pgo.privacy != private_group_privacy::public_group) {
             GOLOS_CHECK_LOGIC(exists &&
@@ -51,7 +55,7 @@ struct private_message_extension_visitor {
 
         if (delete_from != account_name_type() && requester != delete_from) {
             GOLOS_CHECK_LOGIC(is_moder,
-                logic_errors::unauthorized, "You should be moder/admin to delete foreign messages.");
+                logic_errors::unauthorized, "You should be moder to delete foreign messages.");
         }
 
         group = _pgo.group;
@@ -113,6 +117,11 @@ void private_message_evaluator::do_apply(const private_message_operation& op) {
     };
 
     if (id_itr == id_idx.end()) {
+        if (group.size()) {
+            const auto& from = _db.get_account(op.from);
+            check_golos_power(from, _db);
+        }
+
         _db.create<message_object>([&](auto& pmo) {
             from_string(pmo.group, group);
             pmo.from = op.from;
@@ -560,22 +569,10 @@ void private_contact_evaluator::do_apply(const private_contact_operation& op) {
     }
 }
 
-void check_golos_power(const account_object& acc, const database& _db) {
-    auto min_gp = _db.get_min_gp_to_emission();
-    GOLOS_CHECK_LOGIC(acc.vesting_shares >= min_gp.first,
-        logic_errors::too_low_gp,
-        "Too low golos power: ${r} is required, ${p} is present.",
-        ("r", min_gp.second)("p", acc.vesting_shares));
-}
-
 void private_group_evaluator::do_apply(const private_group_operation& op) {
     ASSERT_REQ_HF(STEEMIT_HARDFORK_0_30__236, "private_group_operation");
 
     const auto* pgo = _db.find<private_group_object, by_name>(op.name);
-
-    auto admin = op.admin != account_name_type() ? op.admin : op.creator;
-
-    _db.get_account(admin);
 
     auto now = _db.head_block_time();
 
@@ -587,47 +584,7 @@ void private_group_evaluator::do_apply(const private_group_operation& op) {
             logic_errors::cannot_change_group_encrypted,
             "Cannot make encrypted group not encrypted or vice-versa.");
 
-        auto admins = pgo->admins;
-
         const auto& pgm_idx = _db.get_index<private_group_member_index, by_group_type>();
-
-        auto new_admin = pgm_idx.find(admin);
-
-        if (new_admin == pgm_idx.end() || new_admin->member_type != private_group_member_type::admin) {
-            if (pgo->admins == 1) {
-                auto old_admin = pgm_idx.find(std::make_tuple(pgo->name, private_group_member_type::admin));
-        
-                _db.modify(*old_admin, [&](auto& pgmo) {
-                    pgmo.member_type = private_group_member_type::member;
-                    pgmo.updated = now;
-                });
-
-                --admins;
-            }
-
-            if (new_admin == pgm_idx.end()) {
-                GOLOS_CHECK_LOGIC(admins && admins == MAX_ADMINS_LIMIT,
-                    logic_errors::too_many_admins,
-                    "Too many admins.");
-
-                _db.create<private_group_member_object>([&](auto& pgmo) {
-                    from_string(pgmo.group, op.name);
-                    pgmo.account = admin;
-                    from_string(pgmo.json_metadata, "{}");
-                    pgmo.member_type = private_group_member_type::admin;
-                    pgmo.invited = op.creator;
-                    pgmo.joined = now;
-                    pgmo.updated = now;
-                });
-
-                ++admins;
-            } else {
-                _db.modify(*new_admin, [&](auto& pgmo) {
-                    pgmo.member_type = private_group_member_type::admin;
-                    pgmo.updated = now;
-                });
-            }
-        }
 
         if (pgo->privacy != private_group_privacy::public_group &&
             op.privacy == private_group_privacy::public_group) {
@@ -646,15 +603,19 @@ void private_group_evaluator::do_apply(const private_group_operation& op) {
         _db.modify(*pgo, [&](auto& pgo) {
             from_string(pgo.json_metadata, op.json_metadata);
             pgo.privacy = op.privacy;
-            pgo.admins = admins;
             pgo.pendings = 0;
         });
 
         return;
     }
 
-    const auto& creator = _db.get_account(op.creator);
-    check_golos_power(creator, _db);
+    auto fee = _db.get_witness_schedule_object().median_props.private_group_cost;
+    if (fee.amount != 0) {
+        const auto& creator = _db.get_account(op.creator);
+        GOLOS_CHECK_BALANCE(_db, creator, MAIN_BALANCE, fee);
+        _db.adjust_balance(creator, -fee);
+        _db.adjust_balance(_db.get_account(STEEMIT_WORKER_POOL_ACCOUNT), fee);
+    }
 
     const auto& pgo_idx = _db.get_index<private_group_index, by_owner>();
     auto pgo_itr = pgo_idx.lower_bound(op.creator);
@@ -664,7 +625,7 @@ void private_group_evaluator::do_apply(const private_group_operation& op) {
         if (count == MAX_GROUPS_LIMIT) break;
     }
 
-    GOLOS_CHECK_LOGIC(count && count < MAX_GROUPS_LIMIT,
+    GOLOS_CHECK_LOGIC(count < MAX_GROUPS_LIMIT,
         logic_errors::too_many_groups,
         "Too many groups.");
 
@@ -675,17 +636,6 @@ void private_group_evaluator::do_apply(const private_group_operation& op) {
         pgo.is_encrypted = op.is_encrypted;
         pgo.privacy = op.privacy;
         pgo.created = now;
-        pgo.admins = 1;
-    });
-
-    _db.create<private_group_member_object>([&](auto& pgmo) {
-        from_string(pgmo.group, op.name);
-        pgmo.account = admin;
-        from_string(pgmo.json_metadata, "{}");
-        pgmo.member_type = private_group_member_type::admin;
-        pgmo.invited = op.creator;
-        pgmo.joined = now;
-        pgmo.updated = now;
     });
 }
 
@@ -699,8 +649,11 @@ void private_group_delete_evaluator::do_apply(const private_group_delete_operati
         "Not your group.");
 
     const auto& pgm_idx = _db.get_index<private_group_member_index, by_group_type>();
-    auto pgm_itr = pgm_idx.find(std::make_tuple(op.name, private_group_member_type::member));
-    while (pgm_itr != pgm_idx.end() && pgm_itr->group == pgo.name) {
+    auto pgm_itr = pgm_idx.lower_bound(op.name);
+    while (pgm_itr != pgm_idx.end()) {
+        if (pgm_itr->group != pgo.name) {
+            break;
+        }
         const auto& pgm = *pgm_itr;
         ++pgm_itr;
         _db.remove(pgm);
@@ -712,11 +665,10 @@ void private_group_delete_evaluator::do_apply(const private_group_delete_operati
 void private_group_member_evaluator::do_apply(const private_group_member_operation& op) {
     ASSERT_REQ_HF(STEEMIT_HARDFORK_0_30__236, "private_group_member_operation");
 
-    _db.get_account(op.member);
+    const auto& member = _db.get_account(op.member);
 
     const auto& pgo = _db.get<private_group_object, by_name>(op.name);
 
-    auto admins = pgo.admins;
     auto moders = pgo.moders;
     auto members = pgo.members;
     auto pendings = pgo.pendings;
@@ -727,8 +679,7 @@ void private_group_member_evaluator::do_apply(const private_group_member_operati
     bool requester_exists = requester_itr != pgm_idx.end();
 
     bool is_owner = pgo.owner == op.requester;
-    bool is_admin = is_owner || (requester_exists && requester_itr->member_type == private_group_member_type::admin);
-    bool is_moder = is_admin || (requester_exists && requester_itr->member_type == private_group_member_type::moder);
+    bool is_moder = is_owner || (requester_exists && requester_itr->member_type == private_group_member_type::moder);
     bool is_same = op.requester == op.member;
 
     auto member_itr = pgm_idx.find(std::make_tuple(op.member, op.name));
@@ -744,18 +695,13 @@ void private_group_member_evaluator::do_apply(const private_group_member_operati
         GOLOS_CHECK_LOGIC(pgo.privacy != private_group_privacy::public_group,
             logic_errors::group_is_public, "Group is public.");
         GOLOS_CHECK_LOGIC(is_same, logic_errors::unauthorized, "User can request membership only by theirself.");
+        check_golos_power(member, _db);
         pendings++;
     }
 
     else if (op.member_type == private_group_member_type::moder) {
-        GOLOS_CHECK_LOGIC(is_admin, logic_errors::unauthorized, "Only admins can make members moderators.");
+        GOLOS_CHECK_LOGIC(is_owner, logic_errors::unauthorized, "Only owner can make members moderators.");
         moders++;
-    }
-
-    else if (op.member_type == private_group_member_type::admin) {
-        GOLOS_CHECK_LOGIC(is_owner,
-            logic_errors::unauthorized, "Only owner can make users admins.");
-        admins++;
     }
 
     else if (op.member_type == private_group_member_type::retired) {
@@ -763,16 +709,12 @@ void private_group_member_evaluator::do_apply(const private_group_member_operati
 
         bool is_banned = requester_exists && requester_itr->member_type == private_group_member_type::banned;
         GOLOS_CHECK_LOGIC(!is_banned, logic_errors::unauthorized, "You are banned.");
-
-        bool is_retired = requester_exists && requester_itr->member_type == private_group_member_type::retired;
-        GOLOS_CHECK_LOGIC(!is_retired, logic_errors::unauthorized, "Already retired.");
     }
 
     auto now = _db.head_block_time();
 
     if (member_itr != pgm_idx.end()) {
-        if (op.member_type == private_group_member_type::retired
-                && member_itr->member_type == private_group_member_type::pending) {
+        if (op.member_type == private_group_member_type::retired) {
             pendings--;
             _db.remove(*member_itr);
             return;
@@ -782,8 +724,6 @@ void private_group_member_evaluator::do_apply(const private_group_member_operati
             members--;
         } else if (member_itr->member_type == private_group_member_type::moder) {
             moders--;
-        } else if (member_itr->member_type == private_group_member_type::admin) {
-            admins--;
         } else if (member_itr->member_type == private_group_member_type::pending) {
             pendings--;
         }
@@ -806,7 +746,6 @@ void private_group_member_evaluator::do_apply(const private_group_member_operati
     });
 
     _db.modify(pgo, [&](auto& pgo) {
-        pgo.admins = admins;
         pgo.moders = moders;
         pgo.members = members;
         pgo.pendings = pendings;
